+++ src/common/settings.ts
@@ -8,7 +8,8 @@ export const DIALOG_SIZES: Record<keyof typeof DialogType, [number, number]> = {
 
 export const FEATURES: Record<string, boolean> = {
   IMPORT_DUPLICATE_DETECTION: false,
-}
+} as const
 
+// These are the columns that are used to calculate the hash of a FireTable row
 // ORDER IS IMPORTANT!
 export const HASH_COLUMNS: FireColumn[] = ['iban', 'date', 'amount', 'contra_account', 'description']