+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 4.14.10-beta
+
+- **Refactor** - Refactored `Table` to contain headers now + added a `Pipeline` class which can be used to pipe different stages during import and preview pipelines.
+  - Removed duplicate account RPC functions reducing unnecessary complexity
+  - Removed references to "Banks" instead we see these as "Financial Accounts"
+  - Fixed preview data transport between server and client (still need proper solution)
+
+## 4.14.9-beta
+
+- **Feature** - Support for git worktrees
+
 ## 4.14.8-beta
 
 - **Refactor** - Utilized `Table` and `FireTable` classes extensively in `duplicate-finder.ts`, `account-utils.ts`, `rpc.ts`, and `config.ts` to keep things DRY and SOLID.